• Rename files • Move default batch import into batcher.js • Add file headers
