+++ src/Recoil_index.js
@@ -37,10 +37,7 @@ export type {
   SelectorFamilyOptions,
 } from './recoil_values/Recoil_selectorFamily';
 
-const {unstable_batchedUpdates} = require('./util/ReactBatchedUpdates');
-const {setBatch} = require('./util/Recoil_batch');
-setBatch(unstable_batchedUpdates);
-
+const {batchUpdates, setBatcher} = require('./util/Recoil_batcher');
 const {DefaultValue} = require('./core/Recoil_Node');
 const {RecoilRoot} = require('./core/Recoil_RecoilRoot.react');
 const {isRecoilValue} = require('./core/Recoil_RecoilValue');
@@ -116,5 +113,8 @@ module.exports = {
 
   // Other functions
   isRecoilValue,
-  setBatch,
+
+  // Batching
+  batchUpdates,
+  setBatcher,
 };
