--- src/util/ReactBatchedUpdates.js
@@ -1,2 +1,0 @@
-const {unstable_batchedUpdates} = require('ReactDOM');
-module.exports = {unstable_batchedUpdates};
