--- src/util/ReactBatchedUpdates.native.js
@@ -1,2 +1,0 @@
-const {unstable_batchedUpdates} = require('ReactNative');
-module.exports = {unstable_batchedUpdates};
