+++ src/util/Recoil_ReactBatchedUpdates.native.js
@@ -1,0 +1,16 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @emails oncall+recoil
+ * @flow strict
+ * @format
+ *
+ * This is a stub for some integration into FB internal stuff
+ */
+
+// $FlowExpectedError[cannot-resolve-module]
+const {unstable_batchedUpdates} = require('ReactNative');
+module.exports = {unstable_batchedUpdates};
